Create search.js
NOT YET WORKING. I'm going to need to handle this in-line inside the SearchBar.svelte, then send data to the store. This won't work because we're not mapping to an page endpoint on load. We need to load this data only when searching and that may happen several times without even ever navigating away from the same view. It will probably end up in a helper function.