Update [noc].svelte
Fixes issue with broken back button if navigating from another site. It won't push back to previous page if navigation hasn't been built yet. Because our app is essentially a SPA, when we first navigate to the page, we build our navigation and the back button won't have a preset value.