Dev (#61)
* update with contributing information

* workflows

* update node version

* switch to use npm

* switch to `npm install` to generate required `package-lock`

* attempting test fixes

* test fixes

to avoid installing ts-node, I switched the files over to js.
updated documentation to reflect some of these changes and how to implement testing for future work.

* remove pr approvals from workflow

we don't really need pr approvals in the workflow because github already has this set up in the repo settings directly.

* removed mothballs

one in db and one route mothball.